Create PLPTkinter Test.py
This file is a very basic early work for the first options program GUI, it has one main screen with the option of opening up an additional screen for both the degree path viewing and course prediction. Obviously their functionalities are not implemented yet, but over the next few weeks we can add them and improve the visual design of the program. I will be adding more to this on both Saturday and Sunday.