Half of the base functionality is in.
The program can now display the three main degree paths. It currently is designed that once one is viewed in the pop up window, that window must be closed to view a different one. This step is important to getting the prediction part of the program done, the degree paths are held as dictionaries, for the convenience of viewing and interacting with the program.